django/requirements.txt
Django==2.2
gunicorn==20.0.4
psycopg2==2.8.5
celery==5.1.2
djangorestframework==3.12.4